Добавил класс CSpace, поправил CArray

+++ array.cpp
@@ -2,18 +2,18 @@
 
 namespace mycv {
 
-CArray::CArray(int _size)
+CArray::CArray(size_t _size)
     : m_Size(_size)
 {
-    m_Array = std::make_unique<double[]>(size_t(m_Size));
+    m_Array = std::make_unique<double[]>(m_Size);
 }
 
-int CArray::getSize() const {
+size_t CArray::getSize() const {
     return m_Size;
 }
 
-double& CArray::operator[](int _index) const {
-    return m_Array[size_t(_index)];
+double& CArray::operator[](size_t _index) const {
+    return m_Array[_index];
 }
 
 double* CArray::begin() {
